pyproject.toml
[project]
name = "issuu-downloader-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "img2pdf>=0.6.1",
    "python-telegram-bot>=22.2",
    "requests>=2.32.4",
]
